Extract StudentExceptionAdvices
