+++ Spring-Boot/Spring-Boot-Student/src/main/java/tw/waterball/judgegirl/springboot/student/controllers/StudentController.java
@@ -14,20 +14,13 @@
 package tw.waterball.judgegirl.springboot.student.controllers;
 
 import lombok.AllArgsConstructor;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-import tw.waterball.judgegirl.commons.token.TokenInvalidException;
 import tw.waterball.judgegirl.commons.token.TokenService;
 import tw.waterball.judgegirl.entities.Student;
 import tw.waterball.judgegirl.springboot.student.presenters.GetStudentsPresenter;
 import tw.waterball.judgegirl.springboot.student.presenters.SignInPresenter;
 import tw.waterball.judgegirl.springboot.student.presenters.SignUpPresenter;
 import tw.waterball.judgegirl.springboot.student.view.StudentView;
-import tw.waterball.judgegirl.studentservice.domain.exceptions.DuplicateEmailException;
-import tw.waterball.judgegirl.studentservice.domain.exceptions.StudentEmailNotFoundException;
-import tw.waterball.judgegirl.studentservice.domain.exceptions.StudentIdNotFoundException;
-import tw.waterball.judgegirl.studentservice.domain.exceptions.StudentPasswordIncorrectException;
 import tw.waterball.judgegirl.studentservice.domain.usecases.*;
 
 import java.util.List;
@@ -104,22 +97,6 @@ public class StudentController {
         GetStudentsPresenter presenter = new GetStudentsPresenter();
         getStudentsWithFilterUseCase.execute(new GetStudentsWithFilterUseCase.Request(false, skip, size), presenter);
         return presenter.present();
-    }
-
-
-    @ExceptionHandler({StudentPasswordIncorrectException.class, DuplicateEmailException.class, IllegalArgumentException.class})
-    public ResponseEntity<?> badRequestHandler(Exception err) {
-        return ResponseEntity.badRequest().body(err.getMessage());
-    }
-
-    @ExceptionHandler({StudentIdNotFoundException.class, StudentEmailNotFoundException.class})
-    public ResponseEntity<?> notFoundHandler(Exception err) {
-        return ResponseEntity.notFound().build();
-    }
-
-    @ExceptionHandler({TokenInvalidException.class})
-    public ResponseEntity<?> unauthorizedHandler(Exception err) {
-        return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
     }
 }
